# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi Main API",
        "description": "Tulang punggung dari berbagai aplikasi.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/motorinci/feature-items": {
            "get": {
                "tags": [
                    "Motorinci Feature Items"
                ],
                "summary": "Mendapatkan daftar item fitur motorinci",
                "operationId": "getMotorinciFeatureItemsList",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorinci Feature Items"
                ],
                "summary": "Membuat item fitur motorinci baru",
                "operationId": "storeMotorinciFeatureItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hemat Bahan Bakar"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Sistem injeksi canggih yang membuat konsumsi bahan bakar lebih efisien."
                                    },
                                    "icon": {
                                        "description": "Icon file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/201_Created"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/403_Forbidden"
                    }
                }
            }
        },
        "/api/motorinci/feature-items/{id}": {
            "get": {
                "tags": [
                    "Motorinci Feature Items"
                ],
                "summary": "Mendapatkan satu item fitur motorinci berdasarkan ID",
                "operationId": "getMotorinciFeatureItemById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci feature item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorinci Feature Items"
                ],
                "summary": "Memperbarui item fitur motorinci yang sudah ada",
                "description": "Gunakan method POST dengan _method=PUT di dalam form-data untuk request PUT/PATCH.",
                "operationId": "updateMotorinciFeatureItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci feature item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Kunci Canggih"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Sistem kunci pintar dengan alarm terintegrasi."
                                    },
                                    "icon": {
                                        "description": "New icon file (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/403_Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorinci Feature Items"
                ],
                "summary": "Menghapus item fitur motorinci",
                "operationId": "deleteMotorinciFeatureItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci feature item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204_NoContent"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/403_Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operasi berhasil"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan"
                    }
                },
                "type": "object"
            },
            "ValidationErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data yang diberikan tidak valid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": [
                                "Pesan error."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "200_Success": {
                "description": "Operasi berhasil dieksekusi.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
            "201_Created": {
                "description": "Sumber daya berhasil dibuat.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
            "204_NoContent": {
                "description": "Operasi berhasil tetapi tidak ada konten untuk dikembalikan."
            },
            "401_Unauthorized": {
                "description": "Autentikasi gagal. Token tidak valid atau tidak ada.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "403_Forbidden": {
                "description": "Akses ditolak. Anda tidak memiliki izin.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "404_NotFound": {
                "description": "Sumber daya yang diminta tidak ditemukan.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "422_UnprocessableContent": {
                "description": "Validasi gagal.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorJsonResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Bearer Anda. Contoh: 'Bearer 1|abcdef...'",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}